pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gobiq-knowledge"
version = "0.1.0"
description = "Single source of truth for GoBIQ analytical knowledge — consumed by GoBIQ chat and GoBIQ report."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Proprietary" }
authors = [
    { name = "GoBblecube" },
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]
rag = [
    # Add Milvus / vector store deps here when the rag module is wired up.
    # "pymilvus>=2.3",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
gobiq_knowledge = []

# The knowledge/ tree is shipped as data alongside the package.
# It lives at the repo root rather than under src/ so non-engineer authors
# can find and edit content without spelunking through the Python source.
[tool.setuptools.data-files]
"share/gobiq_knowledge/knowledge" = ["knowledge/**/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
